Move get_version from utils.py to config.py
Motivation: I dont think that it will be used in other places than
config.py because it is used to save specific commit hash in Final variable.
When it was located in the utils.py, it caused a lot of circural imports
warnings because there are many imports from config.py so config.py
should not import from steam_deals module.