Remove unnecessary async defs that were slowing down the API
When we are not using any await and doing maths calculation (not I/O,
but processor bound) we should use def as it creates separate
ThreadPool.

As we used async, fastapi will run the function for each request on the
 same single event thread.

If there is an await for a response, fastapi will use this time for serving
the other requests. And those requests also wait for response from the external
service and hence all of them are served concurrently.

So in this case as most of the time a request to the route waits for response
from some external service and no single request is taking long time doing
something on the single thread they are run, its fine to use async as fastapi
can serve other requests which come at the sametime while some of them are waiting.